requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=3.0.10
mcp>=0.3.0
gunicorn>=20.1.0
websockets>=10.0
requests>=2.27.1